fea: move commit functions to utils and get commit message as arg
